chore: Inline machineRouter.routes into the main router file
